API Gatewayのコードをapi-gatewayワークスペースに移動
