+++ api-gateway/src/internal/mod.rs
