+++ api-gateway/src/main.rs
@@ -1,3 +1,12 @@
-fn main() {
-    println!("Hello, world!");
+mod config;
+mod internal;
+
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    let config = config::Config::new();
+
+    let server = internal::server::Server::new(config);
+    server.run().await;
+
+    Ok(())
 }
