+++ src/main.rs
@@ -1,12 +1,3 @@
-mod config;
-mod internal;
-
-#[tokio::main]
-async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let config = config::Config::new();
-
-    let server = internal::server::Server::new(config);
-    server.run().await;
-
-    Ok(())
+fn main() {
+    println!("Hello, world!");
 }
